+++ storybook/cypress/e2e/document-component.cy.ts
@@ -1,17 +1,37 @@
 import { DocumentComponentMarkFunction } from "@dso-toolkit/core/src/components";
+import { DocumentComponent } from "dso-toolkit";
+import { TemplateResult } from "lit-html";
 
 import { isOdd } from "../support/is-odd";
 
+function setProps(props: Partial<DocumentComponent<TemplateResult>>) {
+  return cy.get("@document-component").then(($el) => {
+    Object.entries(props).forEach(([key, value]) => {
+      cy.wrap($el).invoke("prop", key, value);
+    });
+  });
+}
+
+function expectAlert(text: string | undefined = undefined) {
+  cy.get("@document-component")
+    .shadow()
+    .find("dso-alert.hydrated")
+    .should(text ? "have.text" : "not.exist", text);
+}
+
 describe("Document Component", () => {
+  beforeEach(() => {
+    cy.visit("http://localhost:45000/iframe.html?id=core-document-component--default");
+    cy.get("dso-document-component").as("document-component").should("have.class", "hydrated");
+  });
+
   it("should mark and highlight", () => {
     const marker: DocumentComponentMarkFunction = (text, source) =>
       text
         .split(new RegExp(`(k)`, "gi"))
         .map((item, index) => (isOdd(index) ? { text: item, highlight: source === "kop" && index === 1 } : item));
 
-    cy.visit("http://localhost:45000/iframe.html?id=core-document-component--default")
-      .get("dso-document-component")
-      .should("have.class", "hydrated")
+    cy.get("@document-component")
       .then(($documentComponent) => ($documentComponent[0].mark = cy.spy(marker).as("marker")))
       .get("@marker")
       .invoke("getCalls")
@@ -27,8 +47,7 @@ describe("Document Component", () => {
   });
 
   it("shows a badge with an exclamationmark with tooltip", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-document-component--default")
-      .get("dso-document-component")
+    cy.get("@document-component")
       .invoke("attr", "geneste-ontwerp-informatie", true)
       .invoke("attr", "bevat-ontwerp-informatie", false)
       .shadow()
@@ -48,33 +67,30 @@ describe("Document Component", () => {
         const wijzigactie = wijzigactieState === "default" ? null : wijzigactieState;
         const annotationsWijzigactie = annotationsWijzigactieState === "default" ? null : annotationsWijzigactieState;
 
-        // this test uses args to set the initial state of the component because the argsMapper is needed for the annotations
         cy.visit(
           "http://localhost:45000/iframe.html?id=core-document-component--default&args=open:!true;openAnnotation:!true",
-        )
-          .get("dso-document-component.hydrated")
+        );
+
+        // this test uses args to set the initial state of the component because the argsMapper is needed for the annotations
+        cy.get("@document-component")
           .invoke("prop", "open", true)
           .invoke("prop", "openAnnotation", true)
           .invoke("prop", "wijzigactie", wijzigactie)
           .invoke("prop", "annotationsWijzigactie", annotationsWijzigactie);
 
         if (wijzigactie) {
-          cy.get("dso-document-component.hydrated").should("have.attr", "wijzigactie", wijzigactie);
+          cy.get("@document-component").should("have.attr", "wijzigactie", wijzigactie);
         } else {
-          cy.get("dso-document-component.hydrated").should("not.have.attr", "wijzigactie");
+          cy.get("@document-component").should("not.have.attr", "wijzigactie");
         }
 
         if (annotationsWijzigactie) {
-          cy.get("dso-document-component.hydrated").should(
-            "have.attr",
-            "annotations-wijzigactie",
-            annotationsWijzigactie,
-          );
+          cy.get("@document-component").should("have.attr", "annotations-wijzigactie", annotationsWijzigactie);
         } else {
-          cy.get("dso-document-component.hydrated").should("not.have.attr", "annotations-wijzigactie");
+          cy.get("@document-component").should("not.have.attr", "annotations-wijzigactie");
         }
 
-        cy.get("dso-document-component.hydrated").matchImageSnapshot();
+        cy.get("@document-component").matchImageSnapshot();
       });
     }
   }
@@ -83,23 +99,135 @@ describe("Document Component", () => {
     it(`matches image snapshot ${state} - table-of-contents`, () => {
       const wijzigactie = state === "default" ? null : state;
 
-      cy.visit("http://localhost:45000/iframe.html?id=core-document-component--default")
-        .get("dso-document-component.hydrated")
+      cy.get("@document-component")
         .invoke("prop", "wijzigactie", wijzigactie)
         .invoke("prop", "filtered", false)
-        .invoke("prop", "mode", "table-of-contents");
+        .invoke("prop", "mode", "table-of-contents")
+        .invoke("prop", "gereserveerd", "")
+        .invoke("prop", "vervallen", "");
 
       if (wijzigactie) {
-        cy.get("dso-document-component.hydrated").should("have.attr", "wijzigactie", wijzigactie);
+        cy.get("@document-component").should("have.attr", "wijzigactie", wijzigactie);
       } else {
-        cy.get("dso-document-component.hydrated").should("not.have.attr", "wijzigactie");
+        cy.get("@document-component").should("not.have.attr", "wijzigactie");
       }
 
-      cy.get("dso-document-component.hydrated")
-        .should("have.attr", "mode", "table-of-contents")
-        .and("not.have.attr", "filtered");
+      cy.get("@document-component").should("have.attr", "mode", "table-of-contents").and("not.have.attr", "filtered");
 
-      cy.get("dso-document-component.hydrated").matchImageSnapshot();
+      cy.get("@document-component").matchImageSnapshot();
     });
   }
+
+  it("shows a gereserveerd alert when gereserveerd prop is set and the wijzigactie is 'voegtoe'", () => {
+    setProps({ open: true, gereserveerd: `<Gereserveerd></Gereserveerd>`, vervallen: "", inhoud: "" });
+
+    cy.get("@document-component").shadow().find(".heading-element dso-label").should("have.text", "Gereserveerd");
+
+    expectAlert("Dit onderdeel is gereserveerd voor toekomstige toevoeging.");
+
+    cy.get("@document-component").matchImageSnapshot(`${Cypress.currentTest.title} -- gereserveerd`);
+  });
+
+  it("shows a gereserveerd alert when gereserveerd prop is set", () => {
+    setProps({
+      wijzigactie: null,
+      open: true,
+      gereserveerd: `<Gereserveerd></Gereserveerd>`,
+      vervallen: "",
+      inhoud: "",
+      bevatOntwerpInformatie: false,
+      filtered: false,
+    });
+
+    cy.get("@document-component").shadow().find(".heading-element dso-label").should("have.text", "Gereserveerd");
+
+    expectAlert("Dit onderdeel is gereserveerd voor toekomstige toevoeging.");
+
+    cy.get("@document-component").matchImageSnapshot(`${Cypress.currentTest.title} -- gereserveerd`);
+  });
+
+  it("shows a gereserveerd label with status verwijder when gereserveerd prop is set with wijzigactie='verwijder'", () => {
+    setProps({
+      wijzigactie: null,
+      open: true,
+      filtered: false,
+      gereserveerd: `<Gereserveerd wijzigactie='verwijder'></Gereserveerd>`,
+      vervallen: "",
+      inhoud:
+        "<?xml version='1.0' encoding='UTF-8' standalone='yes'?><Inhoud xmlns='https://standaarden.overheid.nl/stop/imop/tekst/' wijzigactie='voegtoe'><Al>Deze afdeling is van toepassing op gasverbrandingsinstallaties als bedoeld in artikel 6.45 van het Besluit bouwwerken leefomgeving</Al></Inhoud>",
+    });
+
+    cy.get("@document-component")
+      .shadow()
+      .find(".heading-element dso-label")
+      .should("have.text", "Gereserveerd")
+      .should("have.attr", "status", "verwijderd");
+
+    expectAlert();
+
+    cy.get("@document-component").matchImageSnapshot(`${Cypress.currentTest.title} -- gereserveerd verwijderd`);
+  });
+
+  it("shows a gereserveerd label with a verwijderd status (set with wijzigactie='verwijder') and a vervallen label with toegevoegd status (set with wijzigactie='voegtoe')", () => {
+    setProps({
+      wijzigactie: null,
+      open: true,
+      filtered: false,
+      gereserveerd: `<Gereserveerd wijzigactie='verwijder'></Gereserveerd>`,
+      vervallen: `<Vervallen wijzigactie='voegtoe'></Vervallen>`,
+      inhoud: "",
+    });
+
+    cy.get("@document-component")
+      .shadow()
+      .find(".heading-element")
+      .within(() => {
+        cy.get("dso-label[status='verwijderd']").should("exist");
+        cy.get("dso-label[status='toegevoegd']").should("exist");
+      });
+
+    expectAlert();
+
+    cy.get("@document-component").matchImageSnapshot(
+      `${Cypress.currentTest.title} -- gereserveerd verwijderd & vervallen voegtoe`,
+    );
+  });
+
+  it("shows a vervallen label with status toegevoegd when vervallen prop is set with wijzigactie='voegtoe'", () => {
+    setProps({
+      wijzigactie: null,
+      open: true,
+      filtered: false,
+      gereserveerd: "",
+      vervallen: `<Vervallen wijzigactie='voegtoe'></Vervallen>`,
+      inhoud:
+        "<?xml version='1.0' encoding='UTF-8' standalone='yes'?><Inhoud xmlns='https://standaarden.overheid.nl/stop/imop/tekst/' wijzigactie='verwijder'><Al>Deze afdeling is van toepassing op gasverbrandingsinstallaties als bedoeld in artikel 6.45 van het Besluit bouwwerken leefomgeving</Al></Inhoud>",
+    });
+
+    cy.get("@document-component")
+      .shadow()
+      .find(".heading-element dso-label")
+      .should("have.attr", "status", "toegevoegd");
+
+    expectAlert();
+
+    cy.get("@document-component").matchImageSnapshot(`${Cypress.currentTest.title} -- vervallen voegtoe`);
+  });
+
+  it("shows a vervallen label with when vervallen prop is set", () => {
+    setProps({
+      wijzigactie: null,
+      filtered: false,
+      open: true,
+      gereserveerd: "",
+      vervallen: `<Vervallen></Vervallen>`,
+      inhoud: "",
+    });
+
+    cy.get("@document-component").shadow().find(".heading-element dso-label").should("have.text", "Vervallen");
+
+    expectAlert("Dit onderdeel is vervallen.");
+
+    cy.get("@document-component").matchImageSnapshot(`${Cypress.currentTest.title} -- vervallen`);
+  });
 });